+++ src/data_provider/src/sysInfoMac.cpp
@@ -493,7 +493,7 @@ nlohmann::json SysInfo::getGroups() const
 
         groupItem["group_id"] = group["gid"];
         groupItem["group_name"] = (group.contains("groupname") && !group["groupname"].get<std::string>().empty()) ? group["groupname"] : UNKNOWN_VALUE;
-        groupItem["group_description"] = group["comment"];
+        groupItem["group_description"] = (group.contains("comment") && !group["comment"].get<std::string>().empty()) ? group["comment"] : UNKNOWN_VALUE;
         groupItem["group_id_signed"] = group["gid_signed"];
         groupItem["group_uuid"] = UNKNOWN_VALUE;
         groupItem["group_is_hidden"] = group["is_hidden"];